Projeto cassino/requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
psycopg2==2.9.10
sqlparse==0.5.3
tzdata==2024.2
django-cors-headers==3.14.0
djangorestframework-simplejwt==5.2.2
python-dotenv==1.0.0